requirements.txt
streamlit
google-cloud-bigquery
db-dtypes
google-genai
pandas
python-dotenv
fpdf
